feat: swagger for post model

+++ app/Http/Controllers/Post/DestroyController.php
@@ -9,11 +9,19 @@ use Symfony\Component\HttpFoundation\Response;
 class DestroyController extends BaseController
 {
     /**
-     * @param Post $post
+     * @param int $id
      * @return JsonResponse
      */
-    public function __invoke(Post $post): JsonResponse
+    public function __invoke(int $id): JsonResponse
     {
+        $post = Post::find($id);
+
+        if (empty($post)) {
+            return response()->json([
+                "message" => "Post not found."
+            ], Response::HTTP_NOT_FOUND);
+        }
+
         $this->service->destroy($post);
 
         return response()->json([], Response::HTTP_NO_CONTENT);
